+++ types/protocol-mapping.d.ts
@@ -4030,6 +4030,13 @@ export namespace ProtocolMapping {
             returnType: void;
         };
         /**
+         * Registers storage key to be notified when an update occurs to its cache storage list.
+         */
+        'Storage.trackCacheStorageForStorageKey': {
+            paramsType: [Protocol.Storage.TrackCacheStorageForStorageKeyRequest];
+            returnType: void;
+        };
+        /**
          * Registers origin to be notified when an update occurs to its IndexedDB.
          */
         'Storage.trackIndexedDBForOrigin': {
@@ -4048,6 +4055,13 @@ export namespace ProtocolMapping {
          */
         'Storage.untrackCacheStorageForOrigin': {
             paramsType: [Protocol.Storage.UntrackCacheStorageForOriginRequest];
+            returnType: void;
+        };
+        /**
+         * Unregisters storage key from receiving notifications for cache storage.
+         */
+        'Storage.untrackCacheStorageForStorageKey': {
+            paramsType: [Protocol.Storage.UntrackCacheStorageForStorageKeyRequest];
             returnType: void;
         };
         /**